refactor wagtail hooks
