+++ home/tests/test_wagtail_hooks.py
@@ -1,9 +1,7 @@
 import pytest
-from django.test import TestCase
 from wagtail.models import Locale
 
-from home.models import ContentPage, HomePage, OrderedContentSet
-from home.wagtail_hooks import ContentPageAdmin
+from home.models import HomePage, OrderedContentSet
 
 from .page_builder import PageBtn, PageBuilder, WABlk, WABody
 
@@ -15,31 +13,6 @@ def admin_client(client, django_user_model):
     django_user_model.objects.create_superuser(**creds)
     client.login(**creds)
     return client
-
-
-class ContentPageAdminTests(TestCase):
-    def test_body_text_truncation(self):
-        """
-        Body text for web, whatsapp, messenger, and viber, should be truncated in this
-        list view
-        """
-        page = ContentPage(
-            body=[("paragraph", "test " * 100)],
-        )
-        page.whatsapp_body.append(("Whatsapp_Message", {"message": "test " * 100}))
-        page.messenger_body.append(("messenger_block", {"message": "test " * 100}))
-        page.viber_body.append(("viber_message", {"message": "test " * 100}))
-        admin = ContentPageAdmin()
-
-        self.assertEqual(len(admin.web_body(page)), 200)
-        self.assertEqual(len(admin.wa_body(page)), 200)
-        self.assertEqual(len(admin.mess_body(page)), 200)
-        self.assertEqual(len(admin.vib_body(page)), 200)
-        # Web body is different to the rest because of the html
-        self.assertEqual(admin.web_body(page)[-6:], "test …")
-        self.assertEqual(admin.wa_body(page)[-5:], "test…")
-        self.assertEqual(admin.mess_body(page)[-5:], "test…")
-        self.assertEqual(admin.vib_body(page)[-5:], "test…")
 
 
 @pytest.mark.django_db
